refactor: Replace UserDtoRef with UserDto and update related mappings in AuthenticationController, UserController, AuthenticationService, UserService, and UserMapper
