+++ src/main/java/com/regisx001/blog/services/AuthenticationService.java
@@ -1,6 +1,6 @@
 package com.regisx001.blog.services;
 
-import com.regisx001.blog.domain.dto.UserDtoRef;
+import com.regisx001.blog.domain.dto.UserDto;
 import com.regisx001.blog.domain.dto.requests.LoginUserRequest;
 import com.regisx001.blog.domain.dto.requests.VerifyUserRequest;
 import com.regisx001.blog.domain.entities.User;
@@ -8,7 +8,7 @@ import com.regisx001.blog.domain.entities.User;
 public interface AuthenticationService {
     public User register(User user);
 
-    public User authenticate(UserDtoRef.LoginRequest loginUserRequest);
+    public User authenticate(UserDto.LoginRequest loginUserRequest);
 
     public void verifyUser(VerifyUserRequest verifyUserRequest);
