+++ src/main/java/com/regisx001/blog/services/UserService.java
@@ -1,7 +1,7 @@
 package com.regisx001.blog.services;
 
 import com.regisx001.blog.domain.dto.UserDto;
-import com.regisx001.blog.domain.dto.UserDtoRef;
+import com.regisx001.blog.domain.dto.UserDto;
 import com.regisx001.blog.domain.dto.requests.UpdateUserRequest;
 import com.regisx001.blog.domain.entities.User;
 
@@ -13,7 +13,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.web.multipart.MultipartFile;
 
 public interface UserService {
-    public Page<UserDtoRef.Detailed> getAllUsers(Pageable pageable);
+    public Page<UserDto.Detailed> getAllUsers(Pageable pageable);
 
     User findByUsername(String username);
